+++ app/config/locale/sl.php
@@ -1,0 +1,21 @@
+<?php
+
+return [
+    'settings.inspire' => '"Srčika modrosti je umetnost védenja, kaj spregledati."', // This is the line printed in the homepage and console 'view-source'
+    'settings.locale' => 'sl',
+    'settings.direction' => 'ltr',
+
+    // Service - Users
+    'auth.emails.team' => '%s Ekipa',
+    'auth.emails.confirm.title' => 'Potrditev računa',
+    'auth.emails.confirm.body' => 'sl.email.auth.confirm.tpl',
+    'auth.emails.recovery.title' => 'Ponastavitev gesla',
+    'auth.emails.recovery.body' => 'sl.email.auth.recovery.tpl',
+    'auth.emails.invitation.title' => 'Povabilo v %s Ekipo za %s',
+    'auth.emails.invitation.body' => 'sl.email.auth.invitation.tpl',
+
+    'locale.country.unknown' => 'Neznano',
+
+    'countries' => include 'sl.countries.php',
+    'continents' => include 'sl.continents.php',
+];